[transform] move transform tests to new module
