+++ penrose.cabal
@@ -102,6 +102,7 @@ test-suite penrose-testsuite
         Functions.Tests
         Shapes.Tests
         Style.Tests
+        Transforms.Tests
     default-language: Haskell2010
     build-depends:
         base >=4.9 && <4.10,
